pyproject.toml
[tool.poetry]
name = "awesome-blog"
version = "0.1.0"
description = ""
authors = ["Arvind-4 <lordarvindthegamer@gmail.com>"]
readme = "README.md"
packages = [{include = "awesome_blog"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = ">3.2,<3.3"
Pillow = "^9.4.0"
django-ckeditor = "^6.5.1"
python-decouple = "^3.7"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.3.0"


[tool.poetry.group.dev.dependencies]
djlint = "^1.19.11"
black = "^22.12.0"
pylint = "^2.15.10"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
